# SpartanWrath/src/main/java/com/spartanwrath/dto/MembershipDTO.java
package com.spartanwrath.dto;


public record MembershipDTO(String nombre,
                            double precio,
                            String descripcion,
                            Long id,
                            CombatClassDTO combatClass) {}